libhuawei_init: fix up ril for single sim devices.
-Loading the RIL lib for the second sim on single sim devices
causes some crashing by rild. Traditionally we were disabling
"some" of the dual sim props and it was enough but know its
not. So if we are on a single sim device it makes sense not
to load the RIL lib for the second sim and does seem to
resolve ril issues.